fix: Moved semantic timeouts to constants file
